Find all geographies that contain, a certain geography

ods_find_higher_geographies returns a dataset from statistics.gov.scot of all geographies which contain the geography provided

Usage

ods_find_higher_geographies(geography)

Arguments

geography

string. A valid geography unit from statistics.gov.scot.

Value

tibble.

Details

The geography parameter must be passed a valid geography unit.

Examples

ods_find_higher_geographies("S02001257")
#> # A tibble: 3 × 2
#>   geography description       
#>   <chr>     <chr>             
#> 1 S12000033 Council Areas     
#> 2 S08000020 Health Board Areas
#> 3 S92000003 Countries