Return a list of the structure of a dataset

ods_structure returns a list of data about a dataset on statistics.gov.scot

ods_structure(dataset)

Arguments

dataset
string. The identifying final part of a URI for a dataset on statistics.gov.scot.

Value

list. Returns all schemes and concepts within a dataset.

Details

The dataset parameter must be passed a valid dataset name (a full list can be obtained by calling ods_all_datasets().

Examples

ods_structure("average-household-size")

#> $schemes
#>                                                       schemes
#> 1   <http://purl.org/linked-data/sdmx/2009/dimension#refArea>
#> 2 <http://purl.org/linked-data/sdmx/2009/dimension#refPeriod>
#> 3              <http://purl.org/linked-data/cube#measureType>
#> 
#> $categories
#> $categories$refArea
#>  [1] "West Dunbartonshire"   "West Lothian"          "Scotland"             
#>  [4] "Clackmannanshire"      "Dumfries and Galloway" "East Ayrshire"        
#>  [7] "East Lothian"          "East Renfrewshire"     "Na h-Eileanan Siar"   
#> [10] "Falkirk"               "Highland"              "Inverclyde"           
#> [13] "Midlothian"            "Moray"                 "North Ayrshire"       
#> [16] "Orkney Islands"        "Scottish Borders"      "Shetland Islands"     
#> [19] "South Ayrshire"        "South Lanarkshire"     "Stirling"             
#> [22] "Aberdeen City"         "Aberdeenshire"         "Argyll and Bute"      
#> [25] "City of Edinburgh"     "Renfrewshire"          "Angus"                
#> [28] "Dundee City"           "North Lanarkshire"     "East Dunbartonshire"  
#> [31] "Glasgow City"          "Fife"                  "Perth and Kinross"    
#> 
#> $categories$refPeriod
#>  [1] "2003" "2004" "2005" "2006" "2007" "2008" "2009" "2010" "2011" "2012"
#> [11] "2013" "2001" "2002" "2014" "2015" "2016" "2017"
#> 
#> $categories$measureType
#> [1] "Ratio"
#> 
#>